Stopped warnings when feature cannot be split
This was happening if a feature has min and max of same
value, leading to any possible split sending everything to
one side or the other. We then try to fit gaussians, which fails
because there are zero datapoints for one of them. Having a constant
feature doesn't actually indicate anything wrong (could just happen
when we have a few datapoints deep in a tree, especially with bagging)
so I've just stopped the printed warnings (printing screws
concurrency). Hooray.